Final version:
	-- added dull processing flow
	-- added user notifications
